Use sample tezos operations to make the db operations test complete
 * extract data for json decoding and tezos db operations testing into
   separate traits, for easier reference
 * fix small bugs in conversion
